# Add optimized mapreduce implementation for SkipMissing