Based on the AbstractArray methods. Allows the compiler to emit SIMD instructions, for `SkipMissing{Vector{Int}}`, but also improves performance for `SkipMissing{Vector{Float64}}`.

The performance improvement is clear for `Int` (3×-5× less time), but less so for `Float64` (-30%). However, since SIMD now works for `Int`, there's a chance it could eventually be enabled for `Float64`, and it should be quite easier than with the generic `mapreduce` fallback. Also, these 30% are enough to get as fast as R (maybe even a bit faster), which is an important reference point (see [this Discourse thread](https://discourse.julialang.org/t/with-missings-julia-is-slower-than-r/11838)).

Fixes https://github.com/JuliaLang/julia/issues/27679.

```julia
# With Vector{Int}
X1 = rand(Int, 10_000_000);
X2 = Vector{Union{Missing,Int}}(X1);
X3 = ifelse.(rand(length(X2)) .< 0.9, X2, missing);

# With Vector{Float64}
Y1 = rand(Float64, 10_000_000);
Y2 = Vector{Union{Missing,Float64}}(Y1);
Y3 = ifelse.(rand(length(Y2)) .< 0.9, Y2, missing);

using BenchmarkTools

# Before

julia> @btime sum(skipmissing(X1));
 7.316 ms (2 allocations: 32 bytes)

julia> @btime sum(skipmissing(X2));
 18.175 ms (2 allocations: 32 bytes)

julia> @btime sum(skipmissing(X3));
 28.240 ms (2 allocations: 32 bytes)

julia> @btime sum(skipmissing(Y1));
 13.816 ms (2 allocations: 32 bytes)

julia> @btime sum(skipmissing(Y2));
 18.254 ms (2 allocations: 32 bytes)

julia> @btime sum(skipmissing(Y3));
 28.430 ms (2 allocations: 32 bytes)

# After

julia> @btime sum(skipmissing(X1));
  5.918 ms (4 allocations: 64 bytes)

julia> @btime sum(skipmissing(X2));
  6.113 ms (4 allocations: 64 bytes)

julia> @btime sum(skipmissing(X3));
  6.142 ms (4 allocations: 64 bytes)

julia> @btime sum(skipmissing(Y1));
  5.830 ms (4 allocations: 64 bytes)

julia> @btime sum(skipmissing(Y2));
  13.816 ms (4 allocations: 64 bytes)

julia> @btime sum(skipmissing(Y3));
  19.018 ms (4 allocations: 64 bytes)
```