@nanosoldier 2018-06-25
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27743#issuecomment-399878737) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/a4bce79b8a69fe0690f676e15bef5f9baf92e87e/9e7f85d_vs_0978251/report.md). cc @ararslan