@nalimilan 2018-06-26
Nanosoldier results are really contrasted: `sum` is about twice faster when there are 10% of missing values in the input, but about 10%-30% slower when there are no missing values. The pattern is quite consistent across types, so it doesn't look like noise. That's weird, since locally I see only one regression (most likely spurious since it's only for `Int8`).

@ararslan Could something weird be going on with the comparison? I ask because at https://github.com/JuliaLang/julia/pull/27386 there are improvements which I cannot reproduce locally either.