@nalimilan 2018-06-26
Probable explanation here: https://github.com/JuliaLang/julia/pull/27386#issuecomment-400350628. Let's run it again.

@nanosoldier `runbenchmarks("union", vs = ":master")`