@mbauman 2018-06-26
> Probable explanation here: #27386 (comment).

Unfortunately, I don't think that was the case here.  I've not (yet) known Nanosoldier to blatantly lie to us.  In this case, it reported comparing 9e7f85d8639ed16a3859da49cbf93d942f10a247 against 0978251ac5223ffc650c0e8a3c20fb0af83e9d4a.  Those are the correct commits — neither the PR nor master had the random change incorporated.