@nanosoldier 2018-06-26
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27743#issuecomment-400394055) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/36d6f95cdf384005bf7a8d09342c9f4db8cf1a00/d201e62_vs_a23be08/report.md). cc @ararslan