@nalimilan 2018-06-27
Indeed, the regressions are still there. I'm really stuck. I've tried again locally after rebasing on master, and for example I get this for `Vector{Union{Float64,Missing}}` without missing values, while the Nanosoldier report says it's a 32% regression! The increase in allocations is consistent though (and it's negligible).
```julia
julia> master["array"]["(\"skipmissing\", sum, Float64, false)"]
BenchmarkTools.Trial:
  memory estimate:  16 bytes
  allocs estimate:  1
  --------------
  minimum time:     1.462 μs (0.00% GC)
  median time:      1.524 μs (0.00% GC)
  mean time:        1.568 μs (0.00% GC)
  maximum time:     33.038 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> pr["array"]["(\"skipmissing\", sum, Float64, false)"]
BenchmarkTools.Trial:
  memory estimate:  64 bytes
  allocs estimate:  4
  --------------
  minimum time:     732.000 ns (0.00% GC)
  median time:      793.000 ns (0.00% GC)
  mean time:        5.053 μs (0.00% GC)
  maximum time:     42.571 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

What can we do about that? Would somebody check locally whether my results are confirmed?