@nanosoldier 2018-07-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27743#issuecomment-402709408) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/b414bf02dfa5a4616a8374531d8b6ca591a3948d/b9f04a1_vs_ee07b8c/report.md). cc @ararslan