@nalimilan 2018-07-05
I've found out why Nanosoldier find regressions. I'm still not sure why I wasn't able to reproduce the problem before, but at least now I see it when running BaseBenchmarks. It only happens for vectors of ~1000 entries or less (as in the benchmark, but not in my other checks), and only when `!(eltype(A) >: Missing)`. It turns out the benchmarks tested that when there are no missing values, because a comprehension narrowed the type automatically, which I hadn't realized. So I wasn't doing the right comparison.

I've fixed it by only using the new method when `eltype(A) >: Missing`. The specialized method is clearly not supposed to be faster when `!(eltype(A) >: Missing)`  since it adds two loops per block to identify the two first non-missing entries.

There's still a 20% regression locally for `Array{Float32}`, which is probably due to the fact that vectorization currently doesn't work for floats, so the new method doesn't really add much (contrary to integers), and the extra code has a fixed cost which is significant for arrays of ~1000 entries. Let's say whether Nanosoldier confirms it.

@nanosoldier `runbenchmarks("union", vs=":master")`