@nanosoldier 2018-07-05
[Your benchmark job](https://github.com/JuliaLang/julia/pull/27743#issuecomment-402816793) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/ae87f82480242793b483bcbe08e4a454a52de70b/d440db4_vs_d5531c3/report.md). cc @ararslan