@nalimilan 2018-07-05
Good news. Most regressions have disappeared and turned into improvements. The only remaining regressions which are related are for `Float64` and `Float32`. That sounds silly given that it's one of the main use cases for `missing`, but actually they are only there for arrays of length 1000, and progressively disappear for larger vectors, turning into improvements. I've filed https://github.com/JuliaCI/BaseBenchmarks.jl/pull/220 to use larger arrays in the benchmark since that's more representative of work on real data sets. Anyway, this PR is one step in the direction of enabling SIMD instructions, which will make a large difference (as currently for integers).

So barring objections I'll merge this tomorrow.