@matthieugomez 2018-07-07
Awesome! But note that another common operation on datasets is to compute mean per group, where typically length may be < 1000.